# aw04-yangzy0723/server/src/main/java/com/example/webpos/model/Product.java
package com.example.webpos.model;

import jakarta.persistence.*;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Entity
@AllArgsConstructor
@NoArgsConstructor
@Table(name = "products")
public class Product {

    @Id
    @Column(name = "pid")
    @NotEmpty
    private String pid;

    @Column(name = "name")
    @NotEmpty
    private String name;

    @Column(name = "price")
    @NotNull
    private double price;

    @Column(name = "category")
    @NotEmpty
    private String category;

    @Column(name = "img")
    private String img;

    @Column(name = "stock")
    @Min(0)
    private int stock;

    @Column(name = "quantity")
    @Min(0)
    private int quantity;

    // No-arg constructor, all-args constructor, getters and setters are generated by Lombok @Data annotation
}